llm/llm-guard-openai/pyproject.toml
[tool.poetry]
name = "local_model"
version = "0.1.0"
description = ""
authors = ["Vicente Herrera <vicenteherrera@vicenteherrera.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.11.2"
openai = "^1.30.1"
python-dotenv = "^1.0.1"
llm-guard = "^0.3.13"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
mypy = "^1.7.0"
black = "^23.11.0"

[build-system]
requires = ["poetry-core>=1.6.1"]
build-backend = "poetry.core.masonry.api"
